统一提交bug

+++ core/wChannelSocket.cpp
@@ -154,7 +154,7 @@ wStatus wChannelSocket::RecvBytes(char buf[], size_t len, ssize_t *size) {
     return mStatus;
 }
 
-int64_t& wChannelSocket::operator[](uint8_t i) {
+int& wChannelSocket::operator[](uint8_t i) {
     assert(i == 0 || i == 1);
     return mChannel[i];
 }
